+++ src/features/validation/schemaValidation/useSchemaValidation.test.tsx
@@ -4,7 +4,7 @@ import type { JSONSchema7 } from 'json-schema';
 import * as DataModelSchemaProvider from 'src/features/datamodel/DataModelSchemaProvider';
 import * as UseBindingSchema from 'src/features/datamodel/useBindingSchema';
 import { FD } from 'src/features/formData/FormDataWrite';
-import { useSchemaValidation } from 'src/features/validation/frontend/useSchemaValidation';
+import { useSchemaValidation } from 'src/features/validation/schemaValidation/useSchemaValidation';
 import type { IDataType } from 'src/types/shared';
 
 describe('useSchemaValidation', () => {
