+++ src/features/validation/selectors/taskErrors.ts
@@ -10,7 +10,7 @@ import {
   validationsOfSeverity,
 } from 'src/features/validation/utils';
 import { useValidationContext } from 'src/features/validation/validationContext';
-import { getVisibilityForNode } from 'src/features/validation/visibility';
+import { getVisibilityForNode } from 'src/features/validation/visibility/visibilityUtils';
 import { useNodes } from 'src/utils/layout/NodesContext';
 
 /**
