+++ src/features/validation/visibility/visibilityUtils.ts
